clip before conversion in XLS and XLSX
This might speedup by a factor of 90. If e.g. an xlsx has 2 columns
and 41820 rows, it would take 1192 seconds to return a clipped sheet
With pre-clipping, it is just 15 seconds.

If the clipping happens in ParseXLSX, we get down to 13.

Proposal to do the latter also sent to the author of ParseXLSX