Copy files if they cannot be hardlinked
In some cases, hard links are not available.  The `xlsgrep` and `xlscat`
programs are actually the same script and Makefile.PL links them as
such.  Since both items are installed by default in a non-interactive
installation, if the hard link fails, building the module will also
fail.  If for whatever reason we are not able to hard link the files, we
should try just copying them as a backup.

I initially thought using a symbolic link would be a good backup choice
but I am not aware of a portable way to do that.