Use /usr/bin/perl as a standard shebang
Background: for openSUSE we require the scripts and examples to have
the standard path to perl as a shebang. So for this module we
have a patch correcting the custom /pro/bin/perl shebang.

Whenever there are changes to these scripts close to the top the
patch will fail and has to be updated.

I'm not sure what the purpose of the /pro/bin/perl shebang is, but
if it's not necessary this is a kind request to use the standard
path so that we can remove our patch.